requirements.txt
playwright==1.40.0
aiofiles==23.2.1
PyYAML>=6.0
playwright-stealth==1.0.6
